requirements.txt
pandas~=2.3.3
Flask~=3.1.2
requests~=2.32.5
beautifulsoup4~=4.14.2
openpyxl==3.1.5
pillow~=11.3.0